+++ common/persistence/persistenceMetricClients.go
@@ -1187,7 +1187,7 @@ func (p *queuePersistenceClient) EnqueueMessage(message []byte) error {
 	return err
 }
 
-func (p *queuePersistenceClient) ReadMessages(lastMessageID int, maxCount int) ([]*QueueMessage, error) {
+func (p *queuePersistenceClient) ReadMessages(lastMessageID int64, maxCount int) ([]*QueueMessage, error) {
 	p.metricClient.IncCounter(metrics.PersistenceReadQueueMessagesScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceReadQueueMessagesScope, metrics.PersistenceLatency)
@@ -1201,7 +1201,7 @@ func (p *queuePersistenceClient) ReadMessages(lastMessageID int, maxCount int) (
 	return result, err
 }
 
-func (p *queuePersistenceClient) UpdateAckLevel(messageID int, clusterName string) error {
+func (p *queuePersistenceClient) UpdateAckLevel(messageID int64, clusterName string) error {
 	p.metricClient.IncCounter(metrics.PersistenceUpdateAckLevelScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceUpdateAckLevelScope, metrics.PersistenceLatency)
@@ -1215,7 +1215,7 @@ func (p *queuePersistenceClient) UpdateAckLevel(messageID int, clusterName strin
 	return err
 }
 
-func (p *queuePersistenceClient) GetAckLevels() (map[string]int, error) {
+func (p *queuePersistenceClient) GetAckLevels() (map[string]int64, error) {
 	p.metricClient.IncCounter(metrics.PersistenceGetAckLevelScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceGetAckLevelScope, metrics.PersistenceLatency)
@@ -1229,7 +1229,7 @@ func (p *queuePersistenceClient) GetAckLevels() (map[string]int, error) {
 	return result, err
 }
 
-func (p *queuePersistenceClient) DeleteMessagesBefore(messageID int) error {
+func (p *queuePersistenceClient) DeleteMessagesBefore(messageID int64) error {
 	p.metricClient.IncCounter(metrics.PersistenceDeleteQueueMessagesScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceDeleteQueueMessagesScope, metrics.PersistenceLatency)
@@ -1243,7 +1243,7 @@ func (p *queuePersistenceClient) DeleteMessagesBefore(messageID int) error {
 	return err
 }
 
-func (p *queuePersistenceClient) EnqueueMessageToDLQ(message []byte) (int, error) {
+func (p *queuePersistenceClient) EnqueueMessageToDLQ(message []byte) (int64, error) {
 	p.metricClient.IncCounter(metrics.PersistenceEnqueueMessageToDLQScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceEnqueueMessageToDLQScope, metrics.PersistenceLatency)
@@ -1257,7 +1257,7 @@ func (p *queuePersistenceClient) EnqueueMessageToDLQ(message []byte) (int, error
 	return messageID, err
 }
 
-func (p *queuePersistenceClient) ReadMessagesFromDLQ(firstMessageID int, lastMessageID int, pageSize int, pageToken []byte) ([]*QueueMessage, []byte, error) {
+func (p *queuePersistenceClient) ReadMessagesFromDLQ(firstMessageID int64, lastMessageID int64, pageSize int, pageToken []byte) ([]*QueueMessage, []byte, error) {
 	p.metricClient.IncCounter(metrics.PersistenceReadQueueMessagesFromDLQScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceReadQueueMessagesFromDLQScope, metrics.PersistenceLatency)
@@ -1271,7 +1271,7 @@ func (p *queuePersistenceClient) ReadMessagesFromDLQ(firstMessageID int, lastMes
 	return result, token, err
 }
 
-func (p *queuePersistenceClient) DeleteMessageFromDLQ(messageID int) error {
+func (p *queuePersistenceClient) DeleteMessageFromDLQ(messageID int64) error {
 	p.metricClient.IncCounter(metrics.PersistenceDeleteQueueMessageFromDLQScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceDeleteQueueMessageFromDLQScope, metrics.PersistenceLatency)
@@ -1285,7 +1285,7 @@ func (p *queuePersistenceClient) DeleteMessageFromDLQ(messageID int) error {
 	return err
 }
 
-func (p *queuePersistenceClient) RangeDeleteMessagesFromDLQ(firstMessageID int, lastMessageID int) error {
+func (p *queuePersistenceClient) RangeDeleteMessagesFromDLQ(firstMessageID int64, lastMessageID int64) error {
 	p.metricClient.IncCounter(metrics.PersistenceRangeDeleteMessagesFromDLQScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceRangeDeleteMessagesFromDLQScope, metrics.PersistenceLatency)
@@ -1299,7 +1299,7 @@ func (p *queuePersistenceClient) RangeDeleteMessagesFromDLQ(firstMessageID int, 
 	return err
 }
 
-func (p *queuePersistenceClient) UpdateDLQAckLevel(messageID int, clusterName string) error {
+func (p *queuePersistenceClient) UpdateDLQAckLevel(messageID int64, clusterName string) error {
 	p.metricClient.IncCounter(metrics.PersistenceUpdateDLQAckLevelScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceUpdateDLQAckLevelScope, metrics.PersistenceLatency)
@@ -1313,7 +1313,7 @@ func (p *queuePersistenceClient) UpdateDLQAckLevel(messageID int, clusterName st
 	return err
 }
 
-func (p *queuePersistenceClient) GetDLQAckLevels() (map[string]int, error) {
+func (p *queuePersistenceClient) GetDLQAckLevels() (map[string]int64, error) {
 	p.metricClient.IncCounter(metrics.PersistenceGetDLQAckLevelScope, metrics.PersistenceRequests)
 
 	sw := p.metricClient.StartTimer(metrics.PersistenceGetDLQAckLevelScope, metrics.PersistenceLatency)
